+++ src/tech/ml/dataset/base.clj
@@ -792,13 +792,13 @@ This is an interface change and we do apologize!"))))
   - `:key-fn` - function to be applied to column names.  Typical use is:
      `:key-fn keyword`.
   - `:separator` - Add a character separator to the list of separators to auto-detect.
-  - `:csv-parser` - Implementation of univocity's AbstractParser to use.  If not provided
-     a default permissive parser is used.  This way you parse anything that univocity
-     supports (so flat files and such).
+  - `:csv-parser` - Implementation of univocity's AbstractParser to use.  If not
+     provided a default permissive parser is used.  This way you parse anything that
+     univocity supports (so flat files and such).
   - `:bad-row-policy` - One of three options: :skip, :error, :carry-on.  Defaults to
-     :carry-on.  Some csv data has ragged rows and in this case we have several options.
-     If the option is :carry-on then we either create a new column or add missing
-     values for columns that had no data for that row.
+     :carry-on.  Some csv data has ragged rows and in this case we have several
+     options. If the option is :carry-on then we either create a new column or add
+     missing values for columns that had no data for that row.
   - `:skip-bad-rows?` - Legacy option.  Use :bad-row-policy.
   - `:max-chars-per-column` - Defaults to 4096.  Columns with more characters that this
       will result in an exception.
@@ -811,14 +811,14 @@ This is an interface change and we do apologize!"))))
           - Return value must be implement tech.ml.dataset.parser.PColumnParser in
             which case that is used or can return nil in which case the default
             column parser is used.
-    - tuple - pair of [datatype parse-fn] in which case container of type [datatype]
-            will be created.
-            parse-fn can be one of:
+    - tuple - pair of [datatype `parse-data`] in which case container of type
+      [datatype] will be created. `parse-data` can be one of:
         - `:relaxed?` - data will be parsed such that parse failures of the standard
            parse functions do not stop the parsing process.  :unparsed-values and
            :unparsed-indexes are available in the metadata of the column that tell
            you the values that failed to parse and their respective indexes.
-        - `fn?` - function from str-> one of #{:missing :parse-failure value}.
+        - `fn?` - function from str-> one of `:tech.ml.dataset.parser/missing`,
+           `:tech.ml.dataset.parser/parse-failure`, or the parsed value.
            Exceptions here always kill the parse process.  :missing will get marked
            in the missing indexes, and :parse-failure will result in the index being
            added to missing, the unparsed the column's :unparsed-values and
@@ -827,6 +827,9 @@ This is an interface change and we do apologize!"))))
            DateTimeFormatter/ofPattern.
         - `DateTimeFormatter` - use with the appropriate temporal parse static function
            to parse the value.
+        -  map of `{:encode-fn decode-fn} for use only with `:encoded-text` datatype.
+           `encode-fn` must be a transformation from a string to a byte array while
+           `decode-fn` must be a transformation from a byte array to a string.
    - `map?` - the header-name-or-idx is used to lookup value.  If not nil, then
            value can be any of the above options.  Else the default column parser
            is used.