+++ src/tech/ml/dataset/format_sequence.clj
@@ -1,4 +1,4 @@
-(ns tech.ml.dataset.format-sequence
+(ns ^:no-doc tech.ml.dataset.format-sequence
   "This code provided initial by genmeblog after careful consideration
   of R print code"
   (:require [clojure.pprint :refer [cl-format]]))